pyproject.toml
[project]
name = "tigo_python"
version = "0.1.0"
description = "Python client for the Tigo Energy API"
authors = [{ name = "Matt Dreyer", email = "matt_dreyer@hotmail.com" }]
license = "MIT"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "httpx>=0.27.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "pytz>=2023.3",
    "python-dateutil>=2.8.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0", 
    "pytest-cov>=4.1.0",
    "mypy>=1.5.0", 
    "ruff>=0.1.0", 
    "httpx[http2]",
    "python-dotenv>=1.0.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.ruff]
line-length = 88
target-version = "py38"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true